Remove RPC interface, Add stuff to serializer
