+++ serializers/src/commonMain/kotlin/foundation/metaplex/serializers/ByteArrayAsBase64JsonArraySerializer.kt
@@ -1,4 +1,4 @@
-package foundation.metaplex.solanainterfaces
+package foundation.metaplex.serializers
 
 import com.funkatronics.kborsh.BorshDecoder
 import com.funkatronics.kborsh.BorshEncoder
@@ -43,33 +43,3 @@ object ByteArrayAsBase64JsonArraySerializer: KSerializer<ByteArray> {
     }
 }
 
-/**
- * Decodes/Encodes input using the Borsh encoding scheme, and serializes it as a Base64 encoded
- * string, formatted as Json string array:
- * output = {
- *      [
- *          "theBorshEncodedBytesAsBase64String",
- *          "base64"
- *      ]
- * }
- */
-class BorshAsBase64JsonArraySerializer<T>(private val dataSerializer: KSerializer<T>):
-    KSerializer<T?> {
-    private val delegateSerializer = ByteArrayAsBase64JsonArraySerializer
-    override val descriptor: SerialDescriptor = dataSerializer.descriptor
-
-    override fun serialize(encoder: Encoder, value: T?) =
-        encoder.encodeSerializableValue(delegateSerializer,
-            value?.let {
-                BorshEncoder().apply {
-                    encodeSerializableValue(dataSerializer, value)
-                }.borshEncodedBytes
-            } ?: byteArrayOf()
-        )
-
-    override fun deserialize(decoder: Decoder): T? =
-        decoder.decodeSerializableValue(delegateSerializer).run {
-            if (this.isEmpty()) return null
-            BorshDecoder(this).decodeSerializableValue(dataSerializer)
-        }
-}
